include option

+++ dupfilecheck.py
@@ -7,6 +7,7 @@ con=sqlite3.connect('temp.db')
 cur=con.cursor()
 cur.execute("create table temp(name text,md5 text)")
 k = open("log.txt", "w")
+choose=input("같은 파일인지 분석만 하시겠습니까?파일 크기까지 맞을 시 삭제하시겠습니까?(1/2):")
 for (path, dir, files) in os.walk("./"):
     for fn in files:
         pt=os.path.join(path, fn)
@@ -23,8 +24,6 @@ for (path, dir, files) in os.walk("./"):
         cur.execute("insert into temp values(\""+pt+"\",'"+md5+"')")
         con.commit()
         cur.execute("select * from temp where md5='"+md5+"'")
-        #r1=cur.fetchone()
-        #fir=r1[0]
         while(True):
             row=cur.fetchone()
             if row==None:
@@ -34,17 +33,18 @@ for (path, dir, files) in os.walk("./"):
             if(data1==pt and data2==md5):
                 continue
             k.write(pt+"와"+data1+"가 같은 파일인거 같습니다. md5="+data2+"\n")
-            cur.execute("select * from temp where rowid in(select min(rowid) from temp where md5=\""+md5+"\")")
-            fir=cur.fetchone()
-            suvname=str(fir[0])
-            cur.execute("select * from temp where rowid not in(select min(rowid) from temp group by md5)")
-            while(True):
+            if(choose==2):
+                cur.execute("select * from temp where rowid in(select min(rowid) from temp where md5=\""+md5+"\")")
                 fir=cur.fetchone()
-                if fir==None:
-                    break
-                test=str(fir[0])
-                if(os.path.getsize(suvname)==os.path.getsize(test)):
-                    os.remove(test)
+                suvname=str(fir[0])
+                cur.execute("select * from temp where rowid not in(select min(rowid) from temp group by md5)")
+                while(True):
+                    fir=cur.fetchone()
+                    if fir==None:
+                        break
+                    test=str(fir[0])
+                    if(os.path.getsize(suvname)==os.path.getsize(test)):
+                        os.remove(test)
             cur.execute("delete from temp where rowid not in(select min(rowid) from temp group by md5)")
             con.commit()
         print("분석 완료")
